+++ turboui/src/WorkMap/components/TableRow/RowContainer.tsx
@@ -1,4 +1,4 @@
-import classNames from "../../utils/classnames";
+import classNames from "../../../utils/classnames";
 import { useTableRowContext } from "./context";
 
 interface Props {
