+++ turboui/src/WorkMap/components/WorkMapTable/TableHeader.tsx
@@ -1,5 +1,5 @@
 import WorkMap from "..";
-import classNames from "../../utils/classnames";
+import classNames from "../../../utils/classnames";
 
 interface Props {
   filter: WorkMap.Filter;
