+++ turboui/src/WorkMap/stories/WorkMapTable.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, within, userEvent } from "@storybook/test";
-import { WorkMapTable } from "./WorkMapTable";
-import { WorkMap } from ".";
-import { currentYear, currentQuarter, currentMonth } from "../utils/timeframes";
-import { TimeframeSelector } from "../TimeframeSelector";
-import { PrivacyIndicator } from "../PrivacyIndicator";
+import { WorkMapTable } from "../components/WorkMapTable";
+import { WorkMap } from "../components";
+import { currentYear, currentQuarter, currentMonth } from "../../utils/timeframes";
+import { TimeframeSelector } from "../../TimeframeSelector";
+import { PrivacyIndicator } from "../../PrivacyIndicator";
 
 // --- Mock Data ---
 function genAvatar(id: string) {
